requirements.txt
plyvel
ntplib
Twisted
colorlog
simplejson
PyYAML
grpcio-tools
grpcio
google-api-python-client
google-auth
httplib2
service_identity
protobuf
pyopenssl
six
click
pyqrllib==1.2.3
pyqryptonight
pyqrandomx
Flask
json-rpc
idna
cryptography
base58
mock
daemonize